fix types

+++ analytics/analytics/analytic_types/__init__.py
@@ -17,7 +17,7 @@ AnalyticUnitId = str
 ModelCache = dict
 
 # TODO: explicit timestamp / value
-TimeSeries = [List[Tuple[int, int]]]
+TimeSeries = [List[Tuple[int, float]]]
 
 """
 Example:
